[feat] rename screenshot.js to captcha.js

+++ server/captcha.js
@@ -11,7 +11,7 @@ class CaptchaHelper {
     }
 
 
-    async Binarization(img) {
+    async Binarization(img, threshold) {
         let scan = (f) => img.scan(0, 0, img.bitmap.width, img.bitmap.height, (x, y, idx) => {
             const r = img.bitmap.data[idx];
             const g = img.bitmap.data[idx + 1];
@@ -25,17 +25,32 @@ class CaptchaHelper {
 
         img.blur(1);
         scan((r, g, b) => {
-            const t = 108;
-            return (r < t && g < t && b < t) ? this.black : this.white
+            return (r < threshold && g < threshold && b < threshold) ? this.black : this.white
         });
+        // let f = 2;
+        // img.resize({w: img.width * f, h: img.height * f});
+        // img.blur(1);
         return img;
     }
 
     async CropRecognize(img, n, f) {
-        let l = []
+        let l = [];
+        let worker = null;
         for (let i = 0; i < n; i++) {
-            l.push(await f(await img.clone().crop({x: img.width/n * i+2, y: 0, w: img.width/4-4, h: img.height})));
+            let r = await f(
+                await img.clone().crop({x: img.width/n * i+2, y: 0, w: img.width/4-4, h: img.height}),
+                worker
+            )
+            l.push(r.result);
+            worker = r.worker;
         }
+        let str = ` ${l.toString().replaceAll(",", "")} `;
+        let str_len = str.length;
+        console.log(colors("yellow",
+            `====================\n`
+                 + `=${str.padStart(9+str_len/2, " ").padEnd(18, " ")}=\n`
+                 + `====================`));
+        return l;
     }
 
     async RecognizeCaptcha(buffer) {
@@ -48,12 +63,13 @@ class CaptchaHelper {
     }
 
 
-    async RecognizeRotated(img, step, range) {
-        const worker = await createWorker("eng");
+    async RecognizeRotated(img, step, range, worker = null, retworker = false) {
+        worker ??= await createWorker("eng");
 
-
-        let currentCount = 0, current = '';
+        // let currentCount = 0, current = '';
         let maxCount = 0, max = '';
+        
+        let count = {};
 
         for (let i = -range / 2; i <= range / 2; i += step) {
             let {data: {text}}
@@ -61,20 +77,52 @@ class CaptchaHelper {
                 await img.clone().rotate({deg: i}).getBuffer("image/png")
             );
 
-            text = text.replaceAll(/[\s\n\r]/g, "");     // remove whitespace
-            // console.log(colors("yellow", "trial: " + text));
+            // console.log(`origin: ${text}`);
+            text = text.replaceAll(/[^a-z0-9]/gi, "").toUpperCase();     // remove whitespace
+            if (text.length > 1) text = [...new Set(text)].join("");
 
-            if (text.length === 1 && text.match(/[a-z0-9]/i)) {
-                if (text !== current)
-                    [currentCount, current] = [0, text];
-                currentCount++;
-                if (currentCount > maxCount)
-                    [max, maxCount] = [current, currentCount];
+            console.log(colors("grey", `trial: ${text} (max: ${max} * ${maxCount})`));
+
+            if (text.length > 0) {
+                for (const char of text) {
+                    if (count[char] === undefined) count[char] = 0;
+                    count[char]++;
+                    if (count[char] > maxCount) {
+                        max = char;
+                        maxCount = count[char];
+                    }
+                }
             }
         }
-        if(max.match(/[cov]/)) max.toUpperCase();
+
+
+        //     if (text.indexOf(current) < 0 || current === "") {     // text doesn't contain current / not match with current
+        //         console.log(`replace with "${text}", current = "${current}", idx = ${text.indexOf(current)}`);
+        //
+        //         if (current.length > 1) {           // if current contains more than one char
+        //             let text2 = ""
+        //             for (let i = 0; i < current.length; i++) {  // search if text contains on of the char
+        //                 if (text.indexOf(current[i]) >= 0) {    // if text contains this char ?
+        //                     text2 += current[i];                // then add this char
+        //                 }
+        //             }
+        //             text = text2;   // = "" if none of the chars in current match
+        //         }
+        //         // console.log(`replace with "${text}"`);
+        //
+        //         [currentCount, current] = [0, text];        // replace with current
+        //     }
+        //     if (text.length > 0) {
+        //         currentCount++;
+        //         if (currentCount > maxCount)
+        //             [max, maxCount] = [current, currentCount];
+        //     }
+        // }
+
         console.log(colors("yellow", "CAPTCHA RECOGNIZED: " + max));
 
+
+        if (retworker) return { result: max, worker };
 
         await worker.terminate();
         return max;
